refactor(book): move BookData type to types directory and update imports.
