+++ frontend-next/src/types/book.ts
